refactor(core): move metadata table reading out of the file group reader
`FileGroupReader` was doing two jobs. A regular table's file slice is parquet or
lance base files with Avro or parquet log blocks; the metadata table's is HFile
throughout, read by key and merged differently. They shared a struct because
they share the file-slice shape, not because they share a read.

The metadata read moves to `metadata::table::reader`, next to the module that
calls it, and takes its tests and fixture constants with it. `is_metadata_table`
stays behind: it is public API, surfaced through the Python binding, and a
caller choosing a read path has to be able to ask the question.

The instant-range helper is duplicated rather than shared, so neither reader
depends on the other. It is fifteen lines and both sides have tests.

This also removes the HFile imports from the regular reader, which is the point:
the two paths no longer compile against each other.
